function that returns the number of bits you would need to flip to get from one number to another

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -11,9 +11,8 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	int a = n ^ m;
-
-	int count = 0;
+	unsigned long int a = n ^ m;
+	unsigned int count = 0;
 
 	while (a)
 	{
